% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_novices.R
\name{shiny_novices}
\alias{shiny_novices}
\title{Lanceur de l'application rfornoob}
\usage{
shiny_novices()
}
\value{
la fonction lance l'application
}
\description{
Lanceur de l'application rfornoob
}
